Separate GUI and logic in continueAnalysis
